Rename BlocketjobbJobPresenter => Blocketjobb::JobWrapper for consistency
